% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDS2VE.R
\name{showAvailable}
\alias{showAvailable}
\alias{GDS2VE}
\title{ShowAvailable}
\usage{
showAvailable(file, args = c("name", "rowDataColumns", "colDataColumns",
  "infoColumns"))

GDS2VE(file, name = NULL, rowDataColumns = character(),
  colDataColumns = character(), infoColumns = character(),
  rowDataOnDisk = TRUE, colDataOnDisk = TRUE)
}
\arguments{
\item{file}{the path to the gds.class file.}

\item{args}{the arguments in
\code{GDS2VE}.}

\item{name}{the components of the gds file that will be represented
as \code{GDSArray} file.}

\item{rowDataColumns}{which columns of \code{rowData} to
import. The default is ALL.}

\item{colDataColumns}{which columns of \code{colData} to
import. The default is ALL.}

\item{infoColumns}{which columns of \code{infoColumns} to import. The default is
ALL.}

\item{rowDataOnDisk}{whether to save the \code{rowData} as
DelayedArray object. The default is TRUE.}

\item{colDataOnDisk}{whether to save the \code{colData} as
DelayedArray object. The default is TRUE.}
}
\description{
The function to show the available entries for the arguments within
\code{GDS2VE}

Conversion of gds file into SummarizedExperiment.
}
\examples{
\dontrun{
file <- SNPRelate::snpgdsExampleFileName()
se <- GDS2VE(file)
rowData(se)
SummarizedExperiment::colData(se)
metadata(se)
showAvailable(file)
se1 <- GDS2VE(file, rowDataColumns=c("ALLELE"))
SummarizedExperiment::rowRanges(se1)
file <- SeqArray::seqExampleFileName(type="gds")
se <- GDS2VE(file)
names(assays(se))
showAvailable(file)
names <- showAvailable(file, "name")$name
rowdatacols <- showAvailable(file, "rowDataColumns")$rowDataColumns
coldatacols <- showAvailable(file, "colDataColumns")$colDataColumns
infocols <- showAvailable(file, "infoColumns")$infoColumns
se1 <- GDS2VE(
file,
name = names[2],
rowDataColumns = rowdatacols[1:3],
colDataColumns = coldatacols[1],
infoColumns = infocols[c(1,3,5,7)],
rowDataOnDisk = FALSE,
colDataOnDisk = FALSE)
assay(se1)
rowRanges(se1)
colData(se1)
print(object.size(se[TRUE, TRUE]), units="auto")
}
}
