% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-classes.R
\docType{class}
\name{GDSArraySeed-class}
\alias{GDSArraySeed-class}
\title{GDSArraySeed
Generate the seed for gds data for the GDSArray}
\description{
GDSArraySeed
Generate the seed for gds data for the GDSArray
}
