% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-classes.R
\name{GDSArraySeed}
\alias{GDSArraySeed}
\title{GDSAraySeed
The function to generate a GDSArraySeed for the later converting from gds file into GDSARray.}
\usage{
GDSArraySeed(file, name = NA)
}
\arguments{
\item{file}{the gds file name.}

\item{name}{the gds array nodes to be read into GDSArray}
}
\description{
GDSAraySeed
The function to generate a GDSArraySeed for the later converting from gds file into GDSARray.
}
