% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGDSSummarizedExperiment.R
\name{loadGDSSummarizedExperiment}
\alias{loadGDSSummarizedExperiment}
\title{loadGDSSummarizedExperiment
to load the GDS back-end SummarizedExperiment object into R console.}
\usage{
loadGDSSummarizedExperiment(dir = "my_gds_se")
}
\arguments{
\item{dir}{The directory to save the gds format of the array data, and the newly generated SummarizedExperiment object with array data in GDSArray format.}
}
\description{
loadGDSSummarizedExperiment
to load the GDS back-end SummarizedExperiment object into R console.
}
