% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGDSSummarizedExperiment.R
\name{loadVariantExperiment}
\alias{loadVariantExperiment}
\title{loadVariantExperiment to load the GDS back-end SummarizedExperiment
object into R console.}
\usage{
loadVariantExperiment(dir = tempdir())
}
\arguments{
\item{dir}{The directory to save the gds format of the array data,
and the newly generated SummarizedExperiment object with array
data in GDSArray format.}
}
\description{
loadVariantExperiment to load the GDS back-end SummarizedExperiment
object into R console.
}
