% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromGdsfmt.R
\name{makeSummarizedExperimentFromGdsfmt}
\alias{makeSummarizedExperimentFromGdsfmt}
\title{makeSummarizedExperimentFromGdsfmt}
\usage{
makeSummarizedExperimentFromGdsfmt(file, name = NA, frompkg = c("SNPRelate",
  "SeqArray"))
}
\arguments{
\item{file}{the path to the gds.class file.}

\item{name}{the components of the gds file that will be represented as \code{GDSArray} file.}

\item{frompkg}{From which package the gds file is generated.}
}
\description{
Conversion of gds file into SummarizedExperiment
}
