% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVariantExperimentFromSEQGDS.R
\name{makeVariantExperimentFromSEQGDS}
\alias{makeVariantExperimentFromSEQGDS}
\title{makeVariantExperimentFromGDS}
\usage{
makeVariantExperimentFromSEQGDS(
  file,
  assayNames = NULL,
  rowDataColumns = NULL,
  colDataColumns = NULL,
  infoColumns = NULL,
  rowDataOnDisk = TRUE,
  colDataOnDisk = TRUE
)
}
\arguments{
\item{assayNames}{the gds node name that will be read into the
\code{assays} slot and be represented as \code{DelayedArray}
object.}

\item{rowDataColumns}{which columns of \code{rowData} to
import. The default is NULL to read in all variant annotation
info.}

\item{colDataColumns}{which columns of \code{colData} to
import. The default is NULL to read in all sample related
annotation info.}

\item{infoColumns}{which columns of \code{infoColumns} to}

\item{rowDataOnDisk}{whether to save the \code{rowData} as
DelayedArray object. The default is TRUE.}

\item{colDataOnDisk}{whether to save the \code{colData} as
DelayedArray object. The default is TRUE.}
}
\value{
An \code{VariantExperiment} object.
}
\description{
Conversion of seq gds file into SummarizedExperiment.
}
\examples{
file <- SeqArray::seqExampleFileName(type="gds")
se <- makeVariantExperimentFromGDS(file)
## all assay data
names(assays(se))
showAvailable(file)

## only read specific columns for feature / sample annotation. 
assayNamess <- showAvailable(file)$assayNames
rowdatacols <- showAvailable(file)$rowDataColumns
coldatacols <- showAvailable(file)$colDataColumns
infocols <- showAvailable(file)$infoColumns
se1 <- makeVariantExperimentFromGDS(
file,
assayNames = assayNamess[2],
rowDataColumns = rowdatacols[1:3],
colDataColumns = coldatacols[1],
infoColumns = infocols[c(1,3,5,7)],
rowDataOnDisk = FALSE,
colDataOnDisk = FALSE)
assay(se1)

## the rowData(se1) and colData(se1) are now in DataFrame format 
rowData(se1)
colData(se1)
}
