% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVariantExperimentFromSNPGDS.R
\name{makeVariantExperimentFromSNPGDS}
\alias{makeVariantExperimentFromSNPGDS}
\title{makeVariantExperimentFromGDS}
\usage{
makeVariantExperimentFromSNPGDS(
  file,
  assayNames = NULL,
  rowDataColumns = NULL,
  colDataColumns = NULL,
  rowDataOnDisk = TRUE,
  colDataOnDisk = TRUE
)
}
\arguments{
\item{assayNames}{the gds node name that will be read into the
\code{assays} slot and be represented as \code{DelayedArray}
object.}

\item{rowDataColumns}{which columns of \code{rowData} to
import. The default is NULL to read in all variant annotation
info.}

\item{colDataColumns}{which columns of \code{colData} to
import. The default is NULL to read in all sample related
annotation info.}

\item{rowDataOnDisk}{whether to save the \code{rowData} as
DelayedArray object. The default is TRUE.}

\item{colDataOnDisk}{whether to save the \code{colData} as
DelayedArray object. The default is TRUE.}
}
\value{
An \code{VariantExperiment} object.
}
\description{
Conversion of snp gds file into SummarizedExperiment.
}
\examples{
file <- SNPRelate::snpgdsExampleFileName()
se <- makeVariantExperimentFromGDS(file)
rowData(se)
colData(se)
metadata(se)
## Only read specific columns for feature annotation.
showAvailable(file)
se1 <- makeVariantExperimentFromGDS(file, rowDataColumns=c("ALLELE"))
rowRanges(se1)
}
