% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantExperiment-methods.R
\docType{methods}
\name{VariantExperiment-methods}
\alias{VariantExperiment-methods}
\alias{seqAlleleFreq,VariantExperiment-method}
\alias{seqAlleleCount,VariantExperiment-method}
\alias{seqMissing,VariantExperiment-method}
\alias{seqNumAllele,VariantExperiment-method}
\alias{hwe,VariantExperiment-method}
\alias{inbreedCoeff,VariantExperiment-method}
\alias{pca,VariantExperiment-method}
\alias{titv,VariantExperiment-method}
\alias{titv,variantExperiment-method}
\alias{refDosage,VariantExperiment-method}
\alias{altDosage,VariantExperiment-method}
\alias{countSingletons,VariantExperiment-method}
\alias{heterozygosity,VariantExperiment-method}
\alias{homozygosity,VariantExperiment-method}
\alias{meanBySample,VariantExperiment-method}
\alias{isSNV,VariantExperiment-method}
\alias{isVariant,VariantExperiment-method}
\title{Statistical functions for \code{VariantExperiment} objects.}
\usage{
\S4method{seqAlleleFreq}{VariantExperiment}(gdsfile, ref.allele = 0L,
  .progress = FALSE, parallel = seqGetParallel())

\S4method{seqAlleleCount}{VariantExperiment}(gdsfile, ref.allele = 0L,
  .progress = FALSE, parallel = seqGetParallel())

\S4method{seqMissing}{VariantExperiment}(gdsfile, per.variant = TRUE,
  .progress = FALSE, parallel = seqGetParallel())

\S4method{seqNumAllele}{VariantExperiment}(gdsfile)

\S4method{hwe}{VariantExperiment}(gdsobj, permute = FALSE)

\S4method{inbreedCoeff}{VariantExperiment}(gdsobj, margin = c("by.variant",
  "by.sample"), use.names = FALSE)

\S4method{pca}{VariantExperiment}(gdsobj, eigen.cnt = 32)

\S4method{titv}{VariantExperiment}(gdsobj, by.sample = FALSE,
  use.names = FALSE)

\S4method{refDosage}{VariantExperiment}(gdsobj, use.names = TRUE)

\S4method{altDosage}{VariantExperiment}(gdsobj, use.names = TRUE,
  sparse = FALSE)

\S4method{countSingletons}{VariantExperiment}(gdsobj, use.names = FALSE)

\S4method{heterozygosity}{VariantExperiment}(gdsobj, margin = c("by.variant",
  "by.sample"), use.names = FALSE)

\S4method{homozygosity}{VariantExperiment}(gdsobj, allele = c("any", "ref",
  "alt"), margin = c("by.variant", "by.sample"), use.names = FALSE)

\S4method{meanBySample}{VariantExperiment}(gdsobj, var.name,
  use.names = FALSE)

\S4method{isSNV}{VariantExperiment}(gdsobj, biallelic = TRUE)

\S4method{isVariant}{VariantExperiment}(gdsobj, use.names = FALSE)
}
\arguments{
\item{gdsfile}{xx}

\item{ref.allele}{xx}

\item{.progress}{xx}

\item{parallel}{xx}

\item{per.variant}{xx}

\item{gdsobj}{xx}

\item{permute}{xx}

\item{margin}{xx}

\item{use.names}{xx}

\item{eigen.cnt}{xx}

\item{by.sample}{xx}

\item{sparse}{xx}

\item{allele}{xx}

\item{var.name}{xx}

\item{biallelic}{xx}
}
\description{
Statistical functions for \code{VariantExperiment} objects.
}
