% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantExperiment.R
\docType{package}
\name{VariantExperiment}
\alias{VariantExperiment}
\alias{VariantExperiment-package}
\title{VariantExperiment: A package for ...}
\description{
VariantExperiment...

The foo functions ...
}
