% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF2VE.R
\name{VCFtoVE}
\alias{VCFtoVE}
\title{VCFtoVE}
\usage{
VCFtoVE(vcf.fn, out.dir = tempfile(), replace = FALSE,
  info.import = NULL, fmt.import = NULL, sample.info = NULL,
  ignore.chr.prefix = "chr", reference = NULL, start = 1L,
  count = -1L, parallel = FALSE, verbose = FALSE)
}
\arguments{
\item{vcf.fn}{vcf file name. Could be character string or
\code{VcfFile} object or \code{RangedVcfStack} object.}

\item{out.dir}{The directory to save the vcf data, and the newly
generated \code{VariantExperiment} object with array data in
\code{VCFArray} format and annotation data in
\code{elayedDataFrame} format. The default is a temporary
folder.}
}
\description{
\code{VCFtoVE} is the function to convert a vcf file
    into \code{VariantExperiment} object. The genotype data will be
    written as \code{VCFArray} format, which is saved in the
    ‘assays’ slot. The annotation info for variants or samples will
    be written as \code{DelayedDataFrame} object, and saved in the
    \code{rowData} or \code{colData} slot.
}
\examples{
## the vcf file
vcf <- SeqArray::seqExampleFileName("vcf")
ve <- VCFtoVE(vcf)
ve1 <- VCFtoVE(VcfFile(vcf))
all.equal(ve, ve1)
ve0 <- makeSummarizedExperimentFromVCF(vcf)
## RangedVcfStack
extdata <- system.file(package = "GenomicFiles", "extdata")
files <- dir(extdata, pattern="^CEUtrio.*bgz$", full=TRUE)[1:2]
names(files) <- sub(".*_([0-9XY]+).*", "\\\\1", basename(files))
seqinfo <- as(readRDS(file.path(extdata, "seqinfo.rds")), "Seqinfo")
stack <- GenomicFiles::VcfStack(files, seqinfo)
gr <- as(GenomicFiles::seqinfo(stack)[rownames(stack)], "GRanges")
rgstack <- GenomicFiles::RangedVcfStack(stack, rowRanges = gr)
ve2 <- VCFtoVE(rgstack)


}
