% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-class.R,
%   R/makeSummarizedExperimentFromGDS.R
\docType{methods}
\name{gdsfile}
\alias{gdsfile}
\alias{gdsfile,GDSArraySeed-method}
\alias{dim,GDSArraySeed-method}
\alias{dimnames,GDSArraySeed-method}
\alias{gdsfile,GDSArray-method}
\alias{gdsfile,DelayedArray-method}
\alias{GDSArray}
\alias{gdsfile,SummarizedExperiment-method}
\title{GDSArray}
\usage{
gdsfile(x)

\S4method{gdsfile}{GDSArraySeed}(x)

\S4method{dim}{GDSArraySeed}(x)

\S4method{dimnames}{GDSArraySeed}(x)

\S4method{gdsfile}{GDSArray}(x)

\S4method{gdsfile}{DelayedArray}(x)

GDSArray(file, name = NA)

\S4method{gdsfile}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{GDSArray, GDSMatrix, GDSArraySeed or SummarizedExperiment object.}

\item{file}{the gds file name.}

\item{name}{the gds array node to be read into GDSArray}
}
\description{
\code{gdsfile}: on-disk location of GDS file
    represented by this object.

\code{dim}, \code{dimnames}: dimension and dimnames of
    object contained in the GDS file.

\code{GDSArray}: The function to convert a gds file into the GDSArray data structure.
}
