% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-class.R
\name{GDSArray}
\alias{GDSArray}
\title{GDSArray
The function to convert a gds file into the GDSArray data structure.}
\usage{
GDSArray(file, name = NA)
}
\arguments{
\item{file}{the gds file name.}

\item{name}{the gds array nodes to be read into GDSArray}
}
\description{
GDSArray
The function to convert a gds file into the GDSArray data structure.
}
