% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-classes.R, R/GDSArray-methods.R
\docType{methods}
\name{GDSArray}
\alias{GDSArray}
\alias{dim,GDSArraySeed-method}
\alias{dimnames,GDSArraySeed-method}
\alias{gdsfile}
\alias{gdsfile,GDSArraySeed-method}
\alias{gdsfile,GDSArray-method}
\alias{gdsfile,DelayedArray-method}
\alias{gdsfile,SummarizedExperiment-method}
\alias{gdsfile<-}
\title{GDSArray}
\usage{
GDSArray(file, name = NA)

\S4method{dim}{GDSArraySeed}(x)

\S4method{dimnames}{GDSArraySeed}(x)

gdsfile(x)

\S4method{gdsfile}{GDSArraySeed}(x)

\S4method{gdsfile}{GDSArray}(x)

\S4method{gdsfile}{DelayedArray}(x)

\S4method{gdsfile}{SummarizedExperiment}(x)

gdsfile(x) <- value
}
\arguments{
\item{file}{the gds file name.}

\item{name}{the gds array node to be read into GDSArray}

\item{x}{GDSArray, GDSMatrix, GDSArraySeed or SummarizedExperiment object.}
}
\description{
\code{GDSArray}: The function to convert a gds file into the GDSArray data structure.

\code{dim}, \code{dimnames}: dimension and dimnames of
    object contained in the GDS file.

\code{gdsfile}: on-disk location of GDS file
    represented by this object.
}
