% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveVariantExperiment.R
\name{loadVariantExperiment}
\alias{loadVariantExperiment}
\title{loadVariantExperiment to load the GDS back-end SummarizedExperiment
object into R console.}
\usage{
loadVariantExperiment(dir = tempdir())
}
\arguments{
\item{dir}{The directory to save the gds format of the array data,
and the newly generated SummarizedExperiment object with array
data in GDSArray format.}
}
\value{
An \code{VariantExperiment} object.
}
\description{
loadVariantExperiment to load the GDS back-end SummarizedExperiment
object into R console.
}
\examples{
gds <- SeqArray::seqExampleFileName("gds")
ve <- makeVariantExperimentFromGDS(gds)
ve1 <- subsetByOverlaps(ve, GRanges("22:1-48958933"))
aa <- tempfile()
saveVariantExperiment(ve1, dir=aa, replace=TRUE)
loadVariantExperiment(dir = aa)
}
