% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromGDS.R
\name{showAvailable}
\alias{showAvailable}
\alias{makeSummarizedExperimentFromGDS}
\title{ShowAvailable}
\usage{
showAvailable(file, args = c("name", "rowDataColumns", "colDataColumns",
  "infoColumns"))

makeSummarizedExperimentFromGDS(file, name = NULL,
  rowDataColumns = NULL, colDataColumns = NULL, infoColumns = NULL,
  rowDataOnDisk = TRUE, colDataOnDisk = TRUE)
}
\arguments{
\item{file}{the path to the gds.class file.}

\item{args}{the arguments in
\code{makeSummarizedExperimentFromGDS}.}

\item{name}{the components of the gds file that will be represented
as \code{GDSArray} file.}

\item{rowDataColumns}{which columns of \code{rowData} to
import. The default is NULL to read in all variant annotation
info.}

\item{colDataColumns}{which columns of \code{colData} to
import. The default is NULL to read in all sample related
annotation info.}

\item{infoColumns}{which columns of \code{infoColumns} to
import. The default is NULL to read in all info columns.}

\item{rowDataOnDisk}{whether to save the \code{rowData} as
DelayedArray object. The default is TRUE.}

\item{colDataOnDisk}{whether to save the \code{colData} as
DelayedArray object. The default is TRUE.}
}
\description{
The function to show the available entries for the arguments within
\code{makeSummarizedExperimentFromGDS}

Conversion of gds file into SummarizedExperiment.
}
\examples{
## snp gds file
gds <- SNPRelate::snpgdsExampleFileName()
showAvailable(gds)

## sequencing gds file
gds <- SeqArray::seqExampleFileName("gds")
showAvailable(gds)


file <- SNPRelate::snpgdsExampleFileName()
se <- makeSummarizedExperimentFromGDS(file)
rowData(se)
colData(se)
metadata(se)
## Only read specific columns for feature annotation.
showAvailable(file)
se1 <- makeSummarizedExperimentFromGDS(file, rowDataColumns=c("ALLELE"))
SummarizedExperiment::rowRanges(se1)
file <- SeqArray::seqExampleFileName(type="gds")
se <- makeSummarizedExperimentFromGDS(file)
## all assay data
names(assays(se))
showAvailable(file)

## only read specific columns for feature / sample annotation. 
names <- showAvailable(file, "name")$name
rowdatacols <- showAvailable(file, "rowDataColumns")$rowDataColumns
coldatacols <- showAvailable(file, "colDataColumns")$colDataColumns
infocols <- showAvailable(file, "infoColumns")$infoColumns
se1 <- makeSummarizedExperimentFromGDS(
file,
name = names[2],
rowDataColumns = rowdatacols[1:3],
colDataColumns = coldatacols[1],
infoColumns = infocols[c(1,3,5,7)],
rowDataOnDisk = FALSE,
colDataOnDisk = FALSE)
assay(se1)

## the rowData(se1) and colData(se1) are now in DataFrame format 
rowData(se1)
colData(se1)
}
