% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromGDS.R
\name{makeSummarizedExperimentFromGDS}
\alias{makeSummarizedExperimentFromGDS}
\title{makeSummarizedExperimentFromGDS}
\usage{
makeSummarizedExperimentFromGDS(file, name = NA,
  rowDataColumns = character(), colDataColumns = character(),
  colDataOnDisk = TRUE, rowDataOnDisk = TRUE)
}
\arguments{
\item{file}{the path to the gds.class file.}

\item{name}{the components of the gds file that will be represented as \code{GDSArray} file.}

\item{rowDataColumns}{which columns of \code{rowData} to import.}

\item{colDataColumns}{which columns of \code{colData} to import.}

\item{colDataOnDisk}{whether to save the \code{colData} as DelayedArray object. The default is TRUE.}

\item{rowDataOnDisk}{whether to save the \code{rowData} as DelayedArray object. The default is TRUE.}
}
\description{
Conversion of gds file into SummarizedExperiment
}
