% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{showAvailable}
\alias{showAvailable}
\title{ShowAvailable}
\usage{
showAvailable(
  file,
  feature.num,
  sample.num,
  args = c("assayNames", "rowDataColumns", "colDataColumns", "infoColumns")
)
}
\arguments{
\item{file}{the path to the gds.class file.}

\item{feature.num}{the number of features for the gds file. Must be
provided if the file format is not \code{SNP_ARRAY} or
\code{SEQ_ARRAY}.}

\item{sample.num}{the number of features for the gds file. Must be
provided if the file format is not \code{SNP_ARRAY} or
\code{SEQ_ARRAY}.}

\item{args}{the arguments in \code{makeVariantExperimentFromGDS}.}
}
\description{
The function to show the available entries for the arguments within
\code{makeVariantExperimentFromGDS}
}
\examples{
## snp gds file
gds <- SNPRelate::snpgdsExampleFileName()
showAvailable(gds)

## sequencing gds file
gds <- SeqArray::seqExampleFileName("gds")
showAvailable(gds)

}
