% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromGDS.R
\name{showAvailable}
\alias{showAvailable}
\title{ShowAvailable}
\usage{
showAvailable(file, args = c("name", "rowDataColumns", "colDataColumns",
  "infoColumns"))
}
\arguments{
\item{file}{the path to the gds.class file.}

\item{args}{the arguments in \code{makeSummarizedExperimentFromGDS}.}
}
\description{
The function to show the available entries for the arguments within \code{makeSummarizedExperimentFromGDS}
}
