% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantExperiment-class.R
\docType{class}
\name{VariantExperiment-class}
\alias{VariantExperiment-class}
\alias{.VariantExperiment}
\alias{VariantExperiment}
\alias{gdsfile,VariantExperiment-method}
\title{VariantExperiment}
\usage{
VariantExperiment(assays, rowRanges = GRangesList(),
  colData = DelayedDataFrame(), metadata = list())

\S4method{gdsfile}{VariantExperiment}(x)
}
\description{
VariantExperiment could represent big genomic data in RangedSummarizedExperiment object, with on-disk GDS back-end data. The assays are represented by \code{DelayedArray} objects; \code{rowData} and \code{colData} could be represented by \code{DelayedDataFrame} objects.
}
