% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantExperiment-methods.R
\docType{methods}
\name{VariantExperiment-methods}
\alias{VariantExperiment-methods}
\alias{seqAlleleFreq}
\alias{seqAlleleFreq,VariantExperiment-method}
\alias{seqAlleleCount,VariantExperiment-method}
\alias{seqAlleleCount}
\alias{seqMissing,VariantExperiment-method}
\alias{seqMissing}
\alias{seqNumAllele,VariantExperiment-method}
\alias{hwe,VariantExperiment-method}
\alias{inbreedCoeff,VariantExperiment-method}
\alias{pca,VariantExperiment-method}
\alias{titv,VariantExperiment-method}
\alias{titv,variantExperiment-method}
\alias{refDosage,VariantExperiment-method}
\alias{altDosage,VariantExperiment-method}
\alias{countSingletons,VariantExperiment-method}
\alias{heterozygosity,VariantExperiment-method}
\alias{homozygosity,VariantExperiment-method}
\alias{meanBySample,VariantExperiment-method}
\alias{isSNV,VariantExperiment-method}
\alias{isVariant,VariantExperiment-method}
\title{Statistical functions for \code{VariantExperiment} objects.}
\usage{
\S4method{seqAlleleFreq}{VariantExperiment}(gdsfile, ref.allele = 0L,
  .progress = FALSE, parallel = seqGetParallel())

\S4method{seqAlleleCount}{VariantExperiment}(gdsfile, ref.allele = 0L,
  .progress = FALSE, parallel = seqGetParallel())

\S4method{seqMissing}{VariantExperiment}(gdsfile, per.variant = TRUE,
  .progress = FALSE, parallel = seqGetParallel())

\S4method{seqNumAllele}{VariantExperiment}(gdsfile)

\S4method{hwe}{VariantExperiment}(gdsobj, permute = FALSE)

\S4method{inbreedCoeff}{VariantExperiment}(gdsobj,
  margin = c("by.variant", "by.sample"), use.names = FALSE)

\S4method{pca}{VariantExperiment}(gdsobj, eigen.cnt = 32)

\S4method{titv}{VariantExperiment}(gdsobj, by.sample = FALSE,
  use.names = FALSE)

\S4method{refDosage}{VariantExperiment}(gdsobj, use.names = TRUE)

\S4method{altDosage}{VariantExperiment}(gdsobj, use.names = TRUE,
  sparse = FALSE)

\S4method{countSingletons}{VariantExperiment}(gdsobj, use.names = FALSE)

\S4method{heterozygosity}{VariantExperiment}(gdsobj,
  margin = c("by.variant", "by.sample"), use.names = FALSE)

\S4method{homozygosity}{VariantExperiment}(gdsobj, allele = c("any",
  "ref", "alt"), margin = c("by.variant", "by.sample"),
  use.names = FALSE)

\S4method{meanBySample}{VariantExperiment}(gdsobj, var.name,
  use.names = FALSE)

\S4method{isSNV}{VariantExperiment}(gdsobj, biallelic = TRUE)

\S4method{isVariant}{VariantExperiment}(gdsobj, use.names = FALSE)
}
\arguments{
\item{gdsfile}{an \code{VariantExperiment} object that with
synchronized gds file.}

\item{ref.allele}{a single numeric value, a numeric vector or a
character vector; see \code{?SeqArray::seqAlleleFreq} for more
details.}

\item{.progress}{Logical, show process information if \code{TRUE}.}

\item{parallel}{A logical value to indicate serial processing
(\code{FALSE}) or multicore processing (\code{TRUE}). Takes
numeric value or other value; see \code{?SeqArray::seqParallel}
for more details.}

\item{per.variant}{A logical value to indicate whether to calculate
missing rate for variant (\code{TRUE}), or samples
(\code{FALSE}).}

\item{gdsobj}{same as above \code{gdsfile} argument.}

\item{permute}{A logical value indicating whether to permute the
genotypes. See \code{?SeqVarTools::hwe} for more details.}

\item{margin}{"by.variant" OR "by.sample" to indicate
whether the calculation should be done over all samples for
each variant, or over all variants for each sample. See
\code{?SeqVarTools::inbreedCoeff} for more details.}

\item{use.names}{A logical value indicating whether to assign
variant or sample IDs as names of the output vector.}

\item{eigen.cnt}{An integer value indicating how many eigenvalues
and eignvectors to return. The default is 32.}

\item{by.sample}{A logical value indicating whether TiTv should be
calculated by sample or overall for the entire
\code{VariantExperiment} object. See \code{?SeqVarTools::titv}
for more details.}

\item{sparse}{A Logical value indicating whether or not to return
the alterate allele dosage as a sparse matrix. In most cases,
it will dramatically reduce the size of the returned
object. See \code{?SeqVarTools::altDosage} for more details.}

\item{allele}{Choose from "any", "ref," or "alt," to indicate which
alleles to consider when calculating homozygosity. See
\code{?SeqVarTools::homozygosity} for more details.}

\item{var.name}{Character string with name of the variable. Choose
from \code{names(assays(VE_Object))}. See
\code{?SeqVarTools::meanBySample} for more details.}

\item{biallelic}{A logical indicating whether to only consider
biallelic SNVs. See \code{?SeqVarTools::isSNV} for more
details.}
}
\value{
Statistical results in \code{vector} or \code{data.frame} format.
}
\description{
Statistical functions for \code{VariantExperiment} objects.
}
\examples{
gds <- SeqArray::seqExampleFileName("gds")
ve <- makeSummarizedExperimentFromGDS(gds)
ve

## sample missing rate
mr.samp <- seqMissing(ve, per.variant = FALSE)
head(mr.samp)

## hwe
hwe <- hwe(ve)
head(hwe)

## titv ratio by sample / overall
titv <- titv(ve, by.sample=TRUE)
head(titv)
titv(ve, by.sample=FALSE)

## countSingletons
countSingletons(ve)
}
